Refactor: Extract UI component variants and context to separate files
