+++ client/src/components/ui/form-hooks.ts
@@ -1,0 +1,47 @@
+'use client';
+
+import * as React from 'react';
+import { useFormContext } from 'react-hook-form';
+
+type FormFieldContextValue = {
+  name: string;
+};
+
+export const FormFieldContext = React.createContext<FormFieldContextValue>(
+  {} as FormFieldContextValue
+);
+
+type FormItemContextValue = {
+  id: string;
+};
+
+export const FormItemContext = React.createContext<FormItemContextValue>(
+  {} as FormItemContextValue
+);
+
+export const useFormField = () => {
+  const fieldContext = React.useContext(FormFieldContext);
+  const itemContext = React.useContext(FormItemContext);
+  const { getFieldState, formState } = useFormContext();
+
+  // Validate that useFormField is used within proper context
+  if (!fieldContext || !fieldContext.name) {
+    throw new Error('useFormField should be used within <FormField>');
+  }
+
+  if (!itemContext || !itemContext.id) {
+    throw new Error('useFormField should be used within <FormItem>');
+  }
+
+  const fieldState = getFieldState(fieldContext.name, formState);
+  const { id } = itemContext;
+
+  return {
+    id,
+    name: fieldContext.name,
+    formItemId: `${id}-form-item`,
+    formDescriptionId: `${id}-form-item-description`,
+    formMessageId: `${id}-form-item-message`,
+    ...fieldState,
+  };
+};
