+++ client/src/components/ui/form.tsx
@@ -6,7 +6,6 @@ import { Slot } from '@radix-ui/react-slot';
 import {
   Controller,
   FormProvider,
-  useFormContext,
   type ControllerProps,
   type FieldPath,
   type FieldValues,
@@ -14,19 +13,13 @@ import {
 
 import { cn } from '@/lib/utils';
 import { Label } from '@/components/ui/label';
+import {
+  FormFieldContext,
+  FormItemContext,
+  useFormField,
+} from './form-hooks';
 
 const Form = FormProvider;
-
-type FormFieldContextValue<
-  TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
-  name: TName;
-};
-
-const FormFieldContext = React.createContext<FormFieldContextValue>(
-  {} as FormFieldContextValue
-);
 
 const FormField = <
   TFieldValues extends FieldValues = FieldValues,
@@ -40,41 +33,6 @@ const FormField = <
     </FormFieldContext.Provider>
   );
 };
-
-const useFormField = () => {
-  const fieldContext = React.useContext(FormFieldContext);
-  const itemContext = React.useContext(FormItemContext);
-  const { getFieldState, formState } = useFormContext();
-
-  // Validate that useFormField is used within proper context
-  if (!fieldContext || !fieldContext.name) {
-    throw new Error('useFormField should be used within <FormField>');
-  }
-
-  if (!itemContext || !itemContext.id) {
-    throw new Error('useFormField should be used within <FormItem>');
-  }
-
-  const fieldState = getFieldState(fieldContext.name, formState);
-  const { id } = itemContext;
-
-  return {
-    id,
-    name: fieldContext.name,
-    formItemId: `${id}-form-item`,
-    formDescriptionId: `${id}-form-item-description`,
-    formMessageId: `${id}-form-item-message`,
-    ...fieldState,
-  };
-};
-
-type FormItemContextValue = {
-  id: string;
-};
-
-const FormItemContext = React.createContext<FormItemContextValue>(
-  {} as FormItemContextValue
-);
 
 const FormItem = React.forwardRef<
   HTMLDivElement,
@@ -172,7 +130,6 @@ const FormMessage = React.forwardRef<
 FormMessage.displayName = 'FormMessage';
 
 export {
-  useFormField,
   Form,
   FormItem,
   FormLabel,
